Revert DBMS to MySQL
